randomRename.sh: basic functionality
The script will rename a file to a random 10-ASCII-letter string
given the absolute filepath and the absolute path of the parent
directory as the first and second arguments, respectively. Because
KDE services can provide the absolute path of both when calling the
function, it is simple enough to simply require that the parent
directory is given, rather than computed.